+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-* Added alpha arguements to `hist()` and `plot_bands()` (@nkorinek, #409, #410)
+* Changed variable name for `angle_altitude` to `altitude` in `hillshade()` (@nkorinek, )
 
 ## [0.7.3]
 * Added alpha arguments to `hist()` and `plot_bands()` (@nkorinek, #409, #410)